removeLargePairs()
Write a function named removeLargePairs() that accepts as a parameter

a reference to a Vector of integers,
and removes any adjacent pair of integers in the list if the left element of the pair is larger than the right element of the pair. For example, suppose a vector named v stores the following element values:

{1, 1, 2, 2, 4, 3, 6, 5, 3, 7, 8, 5}

We can think of this list as a sequence of pairs:

{{1, 1}, {2, 2}, {4, 3}, {6, 5}, {3, 7}, {8, 5}}

The pairs {4, 3}, {6, 5} and {8, 5} are "large" because the left element is larger than the right one, so these pairs should be removed. So the call of removeLargePairs(v); would change the vector to store:

{1, 1, 2, 2, 3, 7}



If the vector has an odd length, the last element is not part of a pair and is left untouched.
If an empty vector is passed in, the vector should still be empty at the end of the call.
Do not use any other arrays, vectors, or other data structures to help solve this problem, though you can create as many simple variables as you like.